% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fROMS.R
\name{cropROMS}
\alias{cropROMS}
\title{\code{cropROMS} Spatial crop of Norkyst model output}
\usage{
cropROMS(uc = u, vc = v, xdim = c(100:200), ydim = c(1:150))
}
\arguments{
\item{uc}{u current vector data (list with components lon, lat, and u)}

\item{vc}{v current vector data (list with components lon, lat, and v)}

\item{xdim}{Specify subregion (row indexes) of matrix in first dimension.}

\item{ydim}{Specify subregion (column indexes) of matrix in second dimension.}
}
\value{
Returns a cropped version of the dataset, as a list with sublists u and v, whcih are
themselves lists containing objects lat, lon and either u or v
(i.e. similar to unconverted output from readROMS)
}
\description{
\code{cropROMS} Spatial crop of Norkyst model output
}
\details{
Crops matrices and arrays of u and v variables, given specified row and column matrices.
Note! Function currently only functional for current vector (u & v) data combined.
}
\seealso{
\code{\link{readROMS}} to Read Norkyst data,
  \code{\link{imageROMS}} to plot data

Other Norkyst model output: \code{\link{imageROMS}},
  \code{\link{readROMS}}
}
\author{
Martin Biuw
}
\concept{Norkyst model output}
