% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fROMS.R
\name{imageROMS}
\alias{imageROMS}
\title{\code{imageROMS} Plot current vector data from Norkyst model}
\usage{
imageROMS(u = u, v = v, arr.dens = 2, scale = 100,
  arr.col = "slategrey")
}
\arguments{
\item{u}{current vector data (list with components lon, lat, and u)}

\item{v}{current vector data (list with components lon, lat, and v)}

\item{arr.dens}{Plot arrows for every \code{arr.dens} grid cell}

\item{scale}{Scale the length of arrows to provide visually pleasing results}

\item{arr.col}{Colour of arrows}
}
\value{
Silent. Plots the desired current field
}
\description{
\code{imageROMS} Plot current vector data from Norkyst model
}
\details{
Creates an image map, where the colour intensity represents the current speed
and arrows represent direction. This version works only for data in original matrix/array format,
and currently only plots surface currents
}
\seealso{
\code{\link{readROMS}} to Read Norkyst data,
  \code{\link{cropROMS}} to crop Norkyst data

Other Norkyst model output: \code{\link{cropROMS}},
  \code{\link{readROMS}}
}
\author{
Martin Biuw
}
\concept{Norkyst model output}
