% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fROMS.R
\name{readROMS}
\alias{readROMS}
\title{\code{readROMS} Read Norkyst data}
\usage{
readROMS(roms.dir = "R:/Jofrid/Kaldfjord_160m_2014", date = "20141101",
  var = "u", xlim = c(18.3, 18.8), ylim = c(69.6, 69.9),
  to.xyz = F)
}
\arguments{
\item{roms.dir}{Repository of Norkyst data (in NetCDF format)}

\item{date}{Text string matching date info in file name}

\item{var}{Name of variable to read (currently limited to one variable only)}

\item{xlim}{Specify longitude extent of subregion of grid. Note! This is only possible
if data are converted to xyz format}

\item{ylim}{Specify latitude extent of subregion of grid. Note! This is only possible
if data are converted to xyz format}

\item{to.xyz}{Logical, should data be converted to xyz points (TRUE)
or be kept in original array form (FALSE)}
}
\value{
Returns either 1) list containing latitude (matrix), longitude (matrix) and variable (matrix or array)
or 2) data frame of lat/lon/variable data
}
\description{
\code{readROMS} Read Norkyst data
}
\details{
Reads data corresponding to one variable from Norkyst model output covering the Troms coast.
Expects input in NetCDF4 format, and uses the \code{\link{ncdf4}} package.
}
\seealso{
\code{\link{cropROMS}} to crop data in matrix & array form,
  \code{\link{imageROMS}} to plot data

Other Norkyst model output: \code{\link{cropROMS}},
  \code{\link{imageROMS}}
}
\author{
Martin Biuw
}
\concept{Norkyst model output}
